refactor(shortcuts): colocate private components under page wrappers
