+++ src/features/settings/shortcuts/shortcuts-cli/cli-commands-panel/cli-command-row/cli-command-row.helpers.ts
