+++ src/features/settings/shortcuts/shortcuts-cli/cli-commands-panel/cli-commands-panel.tsx
